+++ src/main/java/com/example/team8salecommerce/domain/cart/dto/response/CartResponse.java
@@ -6,6 +6,6 @@ import java.util.List;
 public record CartResponse(
         Long cartId,
         List<CartItemDetailResponse> items,
-        int totalPrice
+        Long totalPrice
 ) {
 }
